pyproject.toml
[tool.poetry]
name = "hw-27"
version = "0.1.0"
description = ""
authors = ["RedKrane86 <redkrane86@gmail.com>"]
readme = "README.md"
packages = [{include = "hw_27"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.1"
psycopg2 = "^2.9.6"
pillow = "^9.5.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
